20-2-2017

+++ application/controllers/Admin.php
@@ -18,6 +18,10 @@ class Admin extends CI_Controller {
 	 * map to /index.php/welcome/<method_name>
 	 * @see https://codeigniter.com/user_guide/general/urls.html
 	 */
+	public function __construct(){
+                parent::__construct();
+                $this->load->model('Model_admin');
+        }
 	public function dashboard($content="",$key=NULL){
 		if($key!=NULL && $key=="xxx!@#xxx"){
 		$data=$content;
@@ -34,16 +38,35 @@ class Admin extends CI_Controller {
 		$this->dashboard($data,"xxx!@#xxx");
 
 	}
-	public function barang()
-	{
-		$data['title']="Barang | Admin";
+	public function liststok($id_barang=NULL){
+		$data['title']="Stok Barang | Admin";
+		$data['logo']="Toko Online";
+		$data['minlogo']="TO";
+		$query=$this->Model_admin->barang_exist_by_id($id_barang);
+		if($id_barang==NULL){
+			$data['barang']=$this->Model_admin->get_stok();
+			$data['content']=$this->load->view("page/liststok",$data,true);
+		}else if($query){
+			$data['editbarang']=$query;
+			$data['content']=$this->load->view("page/liststok",$data,true);
+		}else{
+			$data['barang']=$this->Model_admin->get_stok();
+			$data['content']=$this->load->view("page/liststok",$data,true);
+		}
+		
+		
+		$this->dashboard($data,"xxx!@#xxx");
+	}
+	public function barang(){
+		$data['title']="Tambah Barang | Admin";
 		$data['logo']="Toko Online";
 		$data['minlogo']="TO";
 		$data['content']=$this->load->view("page/barang",$data,true);
 		$this->dashboard($data,"xxx!@#xxx");
 	}
 	function simpan_barang(){
-
+		if($this->input->post('token')!=NULL){
+			$id_barang=$this->Model_admin->get_id_barang_uniq();
 			$datainput=array(
 				'nama_brg'=>$this->input->post('nama_barang'),
 				'kategori'=>implode(",",$this->input->post('kategori_barang')),
@@ -54,19 +77,51 @@ class Admin extends CI_Controller {
 				'harga'=>$this->input->post('harga_barang'),
 				'stock'=>$this->input->post('stok_barang'),
 				'kondisi'=>$this->input->post('kondisi'),
+				'title_seo'=>$this->input->post('title_seo'),
+				'permalink'=>$this->input->post('permalink'),
 				'video'=>$this->input->post('video_barang'),
 				'status'=>$this->input->post('status_barang')
 			);
-			$this->db->set('gambar_1',$this->input->post("file1"));
-			//rename("user/image1.jpg", "user/del/image1.jpg");
-			$this->db->set('gambar_2',$this->input->post("file2"));
-			$this->db->set('gambar_3',$this->input->post("file3"));
-			$this->db->set('gambar_4',$this->input->post("file4"));
-			$this->db->set('gambar_5',$this->input->post("file5"));
-			$this->db->set('gambar_6',$this->input->post("file6"));
-	
-			$this->db->set('date_update', 'NOW()', FALSE);
-			$this->db->set('id_barang', random_string('alnum', 8));
+			$file1=$this->input->post("file1");
+			$file2=$this->input->post("file2");
+			$file3=$this->input->post("file3");
+			$file4=$this->input->post("file4");
+			$file5=$this->input->post("file5");
+			$file6=$this->input->post("file6");
+			$gambar_aktiv=0;
+			if(file_exists("./assets/upload/temp_image/".$file1)&&$file1!=""){
+				$this->db->set('gambar_1',$id_barang."#".$file1);
+				$gambar_aktiv=1;
+				rename("./assets/upload/temp_image/".$file1, "./assets/upload/image/".$id_barang."#".$file1);
+			}
+			if(file_exists("./assets/upload/temp_image/".$file2)&&$file2!=""){
+				$this->db->set('gambar_2',$id_barang."#".$file2);
+				$gambar_aktiv=$gambar_aktiv!=0?$gambar_aktiv:2;
+				rename("./assets/upload/temp_image/".$file2, "./assets/upload/image/".$id_barang."#".$file2);
+			}
+			if(file_exists("./assets/upload/temp_image/".$file3)&&$file3!=""){
+				$this->db->set('gambar_3',$id_barang."#".$file3);
+				$gambar_aktiv=$gambar_aktiv!=0?$gambar_aktiv:3;
+				rename("./assets/upload/temp_image/".$file3, "./assets/upload/image/".$id_barang."#".$file3);
+			}
+			if(file_exists("./assets/upload/temp_image/".$file4)&&$file4!=""){
+				$this->db->set('gambar_4',$id_barang."#".$file4);
+				$gambar_aktiv=$gambar_aktiv!=0?$gambar_aktiv:4;
+				rename("./assets/upload/temp_image/".$file4, "./assets/upload/image/".$id_barang."#".$file4);
+			}
+			if(file_exists("./assets/upload/temp_image/".$file5)&&$file5!=""){
+				$this->db->set('gambar_5',$id_barang."#".$file5);
+				$gambar_aktiv=$gambar_aktiv!=0?$gambar_aktiv:5;
+				rename("./assets/upload/temp_image/".$file5, "./assets/upload/image/".$id_barang."#".$file5);
+			}
+			if(file_exists("./assets/upload/temp_image/".$file6)&&$file6!=""){
+				$this->db->set('gambar_6',$id_barang."#".$file6);
+				$gambar_aktiv=$gambar_aktiv!=0?$gambar_aktiv:6;
+				rename("./assets/upload/temp_image/".$file6, "./assets/upload/image/".$id_barang."#".$file6);
+			}
+			$this->db->set('gambar_aktiv', $gambar_aktiv);
+			$this->db->set('date_update', 'NOW()', FALSE); 
+			$this->db->set('id_barang', $id_barang);
 			$this->db->insert("barang",$datainput);
 			if($this->db->affected_rows()>0){
 				$this->session->set_flashdata("simpan",array("status"=>true));
@@ -74,14 +129,20 @@ class Admin extends CI_Controller {
 				$this->session->set_flashdata("simpan",array("status"=>false));
 			}
 			redirect("admin/barang");
+		}else{
+			show_404();
+		}
 	}
-	public function upload_image(){
+	public function upload_image($edit=0){
 		if($this->input->is_ajax_request()){
-			$config['upload_path'] = './assets/upload/temp_image/';
+			if($edit==1){$folder="image";}
+			else{$folder="temp_image";}
+			$config['upload_path'] = './assets/upload/'.$folder.'/';
 			$config['allowed_types'] = 'jpg|png';
 			$config['max_size']     = '500';
 			$config['max_width'] = '768';
 			$config['max_height'] = '768';
+			$config['file_ext_tolower'] = TRUE;
 			$config['file_name'] = "1_".$_FILES['file']['name'];
 			$this->load->library('upload', $config);
 			$this->upload->initialize($config);
@@ -94,16 +155,22 @@ class Admin extends CI_Controller {
 	}
 	public function remove_image(){
 		if($this->input->is_ajax_request()){
-		   $storeFolder = './assets/upload/temp_image/';   //2
-		   $file=$this->input->post("filename");
-		   $dataimage=json_decode($file);
-		   $file=$dataimage->message;	
+			if($this->input->post('edit')=="1"){
+				$folder="image/";
+			}else{
+				$folder="temp_image/";
+			}
+		   $storeFolder = './assets/upload/'.$folder;   //2
+		   $file=$this->input->post("file");
+		   if(file_exists($storeFolder.$file)){	
 		   	if(!unlink($storeFolder.$file)){
-		   		echo "gagal".$storeFolder.$file;
+		   		echo json_encode(array("status"=>false,"message"=>"File not exist"));
 		   	}else{
-		   		echo "berhasil";
+		   		echo json_encode(array("status"=>true,"message"=>$file));
 		   	}
-		   
+		   }else{
+		   	echo json_encode(array("status"=>false,"message"=>"File not exist"));
+		   }
 		}
 	}
 	public function not_found(){
